Only delete cache file if actually exists

+++ hloc/match_features.py
@@ -280,7 +280,7 @@ def match_from_paths(
             logger.info(f'Matching interrupted. Started writing pairs cache from {idx}')
             pairs_cache_path.write_text("\n".join(" ".join([i, j]) for i, j in dataset.pairs[idx + 1:]))
             logger.info('Finished writing pairs cache')
-        else:
+        elif pairs_cache_path.exists():
             pairs_cache_path.unlink()
             logger.info('Finished matching, deleting pairs cache')
     logger.info('Finished exporting matches.')
